+++ CHANGELOG.md
@@ -6,8 +6,9 @@
 * Check if file exists at beginning of reveal command.
 * Proper error message if editor is not found in PATH.
 * Fix wrong export when a single resource was targeted
-* Prefer folder "public-keys" over working dir if present
+* Prefer folder "public-keys" over working dir if present when looking for public keys
 * Increase specifity of blacklisted key regexes
+* Prefer folder <namespace> over working dir if present when looking for param files
 
 ## 0.3.0